support the tag include: new tagKV , with_precise and witout_tag option ,  tag=*
influxdb not support SQLSessionIT in scaleout

influxdb not support SQLSessionIT in scaleout

influxdb not support SQLSessionIT in scaleout

influxdb not support SQLSessionIT in scaleout

SQLSessionIT not right in influxdb

SQLSessionIT not right in influxdb

SQLSessionIT not right in influxdb

rename the action var

fix it

off the debug

remove some property

test the influxdb tag ok? and BaseSessionIT for DB-CE

test the influxdb tag ok? and BaseSessionIT for DB-CE 2

test the influxdb tag ok? and BaseSessionIT for DB-CE 3

debug influxdb

debug influxdb